Add API documentation, saperate out cluster logic in another file
