+++ src/app/clusters/wifi-network-diagnostics-server/BUILD.gn
@@ -17,6 +17,8 @@ import("//build_overrides/chip.gni")
 source_set("wifi-network-diagnostics-server") {
   sources = [
     "WiFiDiagnosticsTestEventTriggerHandler.h",
+    "wifi-network-diagnostics-logic.cpp",
+    "wifi-network-diagnostics-logic.h",
     "wifi-network-diagnostics-cluster.cpp",
     "wifi-network-diagnostics-cluster.h",
   ]
