+++ src/app/clusters/wifi-network-diagnostics-server/tests/TestWiFiNetworkDiagnosticsCluster.cpp
@@ -52,7 +52,7 @@ TEST_F(TestWiFiNetworkDiagnosticsCluster, CompileTest)
     };
 
     NullProvider nullProvider;
-    WiFiDiagnosticsServer cluster(kRootEndpointId, nullProvider, enabledAttributes, BitFlags<WiFiNetworkDiagnostics::Feature>(0));
+    WiFiDiagnosticsServerCluster cluster(kRootEndpointId, nullProvider, enabledAttributes, BitFlags<WiFiNetworkDiagnostics::Feature>(0));
 
     // Essentially say "code executes"
     ASSERT_EQ(cluster.GetClusterFlags({ kRootEndpointId, WiFiNetworkDiagnostics::Id }), BitFlags<ClusterQualityFlags>());
@@ -70,7 +70,7 @@ TEST_F(TestWiFiNetworkDiagnosticsCluster, AttributesTest)
         };
 
         NullProvider nullProvider;
-        WiFiDiagnosticsServer cluster(kRootEndpointId, nullProvider, enabledAttributes,
+        WiFiDiagnosticsServerCluster cluster(kRootEndpointId, nullProvider, enabledAttributes,
                                       BitFlags<WiFiNetworkDiagnostics::Feature>(0));
 
         // without any enabled attributes, no commands are accepted
@@ -117,7 +117,7 @@ TEST_F(TestWiFiNetworkDiagnosticsCluster, AttributesTest)
         };
 
         ResetCountsProvider resetCountsProvider;
-        WiFiDiagnosticsServer cluster(kRootEndpointId, resetCountsProvider, enabledAttributes,
+        WiFiDiagnosticsServerCluster cluster(kRootEndpointId, resetCountsProvider, enabledAttributes,
                                       BitFlags<WiFiNetworkDiagnostics::Feature>(WiFiNetworkDiagnostics::Feature::kErrorCounts));
 
         ReadOnlyBufferBuilder<DataModel::AcceptedCommandEntry> commandsBuilder;
@@ -223,7 +223,7 @@ TEST_F(TestWiFiNetworkDiagnosticsCluster, AttributesTest)
         BitFlags<WiFiNetworkDiagnostics::Feature> features;
         features.Set(WiFiNetworkDiagnostics::Feature::kErrorCounts);
         features.Set(WiFiNetworkDiagnostics::Feature::kPacketCounts);
-        WiFiDiagnosticsServer cluster(kRootEndpointId, allProvider, enabledAttributes, features);
+        WiFiDiagnosticsServerCluster cluster(kRootEndpointId, allProvider, enabledAttributes, features);
 
         ReadOnlyBufferBuilder<DataModel::AcceptedCommandEntry> commandsBuilder;
         ASSERT_EQ(cluster.AcceptedCommands(ConcreteClusterPath(kRootEndpointId, WiFiNetworkDiagnostics::Id), commandsBuilder),
