+++ src/app/clusters/wifi-network-diagnostics-server/wifi-network-diagnostics-logic.cpp
@@ -1,0 +1,179 @@
+/**
+ *
+ *    Copyright (c) 2025 Project CHIP Authors
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
+#include <app/clusters/wifi-network-diagnostics-server/wifi-network-diagnostics-logic.h>
+#include <app/EventLogging.h>
+#include <clusters/WiFiNetworkDiagnostics/Attributes.h>
+#include <clusters/WiFiNetworkDiagnostics/Ids.h>
+#include <lib/support/logging/CHIPLogging.h>
+#include <tracing/macros.h>
+#include <tracing/metric_event.h>
+
+using namespace chip;
+using namespace chip::app;
+using namespace chip::app::Clusters;
+using namespace chip::app::Clusters::WiFiNetworkDiagnostics;
+using chip::DeviceLayer::DiagnosticDataProvider;
+using chip::DeviceLayer::GetDiagnosticDataProvider;
+
+namespace chip {
+namespace app {
+namespace Clusters {
+
+CHIP_ERROR WiFiDiagnosticsServerLogic::ReadWiFiBssId(AttributeValueEncoder & aEncoder)
+{
+    Attributes::Bssid::TypeInfo::Type bssid;
+
+    uint8_t bssidBytes[chip::DeviceLayer::kMaxHardwareAddrSize];
+    MutableByteSpan bssidSpan(bssidBytes);
+    if (mDiagnosticProvider.GetWiFiBssId(bssidSpan) == CHIP_NO_ERROR)
+    {
+        if (!bssidSpan.empty())
+        {
+            bssid.SetNonNull(bssidSpan);
+            ChipLogProgress(Zcl, "Node is currently connected to Wi-Fi network with BSSID:");
+            ChipLogByteSpan(Zcl, bssidSpan);
+        }
+    }
+    else
+    {
+        ChipLogProgress(Zcl, "The WiFi interface is not currently connected.");
+    }
+
+    return aEncoder.Encode(bssid);
+}
+
+CHIP_ERROR WiFiDiagnosticsServerLogic::ReadSecurityType(AttributeValueEncoder & aEncoder)
+{
+    Attributes::SecurityType::TypeInfo::Type securityType;
+    SecurityTypeEnum value = SecurityTypeEnum::kUnspecified;
+
+    if (mDiagnosticProvider.GetWiFiSecurityType(value) == CHIP_NO_ERROR)
+    {
+        securityType.SetNonNull(value);
+        ChipLogProgress(Zcl, "The current type of Wi-Fi security used: %d", to_underlying(value));
+    }
+    else
+    {
+        ChipLogProgress(Zcl, "The WiFi interface is not currently configured or operational.");
+    }
+
+    return aEncoder.Encode(securityType);
+}
+
+CHIP_ERROR WiFiDiagnosticsServerLogic::ReadWiFiVersion(AttributeValueEncoder & aEncoder)
+{
+    Attributes::WiFiVersion::TypeInfo::Type version;
+    WiFiVersionEnum value = WiFiVersionEnum::kUnknownEnumValue;
+
+    if (mDiagnosticProvider.GetWiFiVersion(value) == CHIP_NO_ERROR)
+    {
+        version.SetNonNull(value);
+        ChipLogProgress(Zcl, "The current 802.11 standard version in use by the Node: %d", to_underlying(value));
+    }
+    else
+    {
+        ChipLogProgress(Zcl, "The current 802.11 standard version in use by the Node is not available");
+    }
+
+    return aEncoder.Encode(version);
+}
+
+CHIP_ERROR WiFiDiagnosticsServerLogic::ReadChannelNumber(AttributeValueEncoder & aEncoder)
+{
+    Attributes::ChannelNumber::TypeInfo::Type channelNumber;
+    uint16_t value = 0;
+
+    if (mDiagnosticProvider.GetWiFiChannelNumber(value) == CHIP_NO_ERROR)
+    {
+        channelNumber.SetNonNull(value);
+        ChipLogProgress(Zcl, "The channel that Wi-Fi communication is currently operating on is: %d", value);
+    }
+    else
+    {
+        ChipLogProgress(Zcl, "The WiFi interface is not currently configured or operational.");
+    }
+
+    return aEncoder.Encode(channelNumber);
+}
+
+CHIP_ERROR WiFiDiagnosticsServerLogic::ReadWiFiRssi(AttributeValueEncoder & aEncoder)
+{
+    Attributes::Rssi::TypeInfo::Type rssi;
+    int8_t value = 0;
+
+    if (mDiagnosticProvider.GetWiFiRssi(value) == CHIP_NO_ERROR)
+    {
+        rssi.SetNonNull(value);
+        ChipLogProgress(Zcl, "The current RSSI of the Node's Wi-Fi radio in dB: %d", value);
+        MATTER_LOG_METRIC(chip::Tracing::kMetricWiFiRSSI, value);
+    }
+    else
+    {
+        ChipLogProgress(Zcl, "The WiFi interface is not currently configured or operational.");
+    }
+
+    return aEncoder.Encode(rssi);
+}
+
+void WiFiDiagnosticsServerLogic::OnDisconnectionDetected(uint16_t reasonCode)
+{
+    MATTER_TRACE_SCOPE("OnDisconnectionDetected", "WiFiDiagnosticsDelegate");
+    ChipLogProgress(Zcl, "WiFiDiagnosticsDelegate: OnDisconnectionDetected");
+
+    Events::Disconnection::Type event{ reasonCode };
+    EventNumber eventNumber;
+
+    if (CHIP_NO_ERROR != LogEvent(event, kRootEndpointId, eventNumber))
+    {
+        ChipLogError(Zcl, "WiFiDiagnosticsDelegate: Failed to record Disconnection event");
+    }
+}
+
+void WiFiDiagnosticsServerLogic::OnAssociationFailureDetected(uint8_t associationFailureCause, uint16_t status)
+{
+    MATTER_TRACE_SCOPE("OnAssociationFailureDetected", "WiFiDiagnosticsDelegate");
+    ChipLogProgress(Zcl, "WiFiDiagnosticsDelegate: OnAssociationFailureDetected");
+
+    Events::AssociationFailure::Type event{ static_cast<AssociationFailureCauseEnum>(associationFailureCause), status };
+
+    EventNumber eventNumber;
+
+    if (CHIP_NO_ERROR != LogEvent(event, kRootEndpointId, eventNumber))
+    {
+        ChipLogError(Zcl, "WiFiDiagnosticsDelegate: Failed to record AssociationFailure event");
+    }
+}
+
+void WiFiDiagnosticsServerLogic::OnConnectionStatusChanged(uint8_t connectionStatus)
+{
+    MATTER_TRACE_SCOPE("OnConnectionStatusChanged", "WiFiDiagnosticsDelegate");
+    ChipLogProgress(Zcl, "WiFiDiagnosticsDelegate: OnConnectionStatusChanged");
+
+    Events::ConnectionStatus::Type event{ static_cast<ConnectionStatusEnum>(connectionStatus) };
+
+    EventNumber eventNumber;
+
+    if (CHIP_NO_ERROR != LogEvent(event, kRootEndpointId, eventNumber))
+    {
+        ChipLogError(Zcl, "WiFiDiagnosticsDelegate: Failed to record ConnectionStatus event");
+    }
+}
+
+} // namespace Clusters
+} // namespace app
+} // namespace chip
